pyproject.toml
[project]
name = "pr-autocot-experiment"
version = "0.1.0"
description = "Paraphrase-Robust Auto-CoT experiment for chain-of-thought demonstration construction"
requires-python = ">=3.9"
dependencies = [
    # Core experiment framework
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    
    # Data and ML
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    
    # LLM API
    "openai>=1.0.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
